requirements.txt
python-dotenv
certspy
langchain_openai
langchain
ipwhois
